factor out redis connection to global settings
