+++ .gitignore
@@ -1,6 +1,8 @@
 data
 worker_data
+build
 *.pyc
 *.swp
+*.log
 /*.egg-info
 /*.rdb
