+++ scripts/job_creator.py
@@ -9,14 +9,13 @@ from modelrunner import config
 from tornado.options import parse_command_line, parse_config_file
 
 import modelrunner as mr
+import modelrunner.settings as settings
 
 # so we can load config via cmd line args
 parse_command_line()
 parse_config_file(config.options.config_file)
 
-
-jm = mr.JobManager(config.options.redis_url,
-                   config.options.primary_url,
+jm = mr.JobManager(config.options.primary_url,
                    config.options.worker_url,
                    config.options.data_dir,
                    {})
