+++ scripts/job_migration_0_4_0.py
@@ -13,6 +13,7 @@ import urllib2
 import logging
 from modelrunner import config
 import modelrunner as mr
+import modelrunner.settings as settings
 
 from tornado.options import parse_command_line, parse_config_file
 
@@ -46,14 +47,16 @@ def test_url(url):
 # get the command_ keys
 command_dict = config.options.group_dict("model_command")
 
-jm = mr.JobManager(config.options.redis_url,
-                   config.options.primary_url,
+# initialize the global application settings
+settings.init_redis_connection(config.options.redis_url)
+
+jm = mr.JobManager(config.options.primary_url,
                    config.options.worker_url,
                    config.options.data_dir,
                    command_dict,
                    config.options.worker_is_primary)
 
-jobs = jm.get_jobs()
+obs = jm.get_jobs()
 
 # get the log file for all jobs
 for job in jobs:
@@ -68,5 +71,4 @@ for job in jobs:
         jm.add_update_job_table(job)
         # push message to primary to get data for job
         primary_queue = "modelrunner:queues:" + jm.primary_url
-        jm.rdb.rpush(primary_queue, job.uuid)
-            
+        settings.redis_connection.rpush(primary_queue, job.uuid)
