+++ scripts/job_worker.py
@@ -16,6 +16,7 @@ from modelrunner import config
 from tornado.options import parse_command_line, parse_config_file
 
 import modelrunner as mr
+import modelrunner.settings as settings
 
 # setup log
 logger = logging.getLogger('modelrunner')
@@ -32,8 +33,10 @@ parse_config_file(config.options.config_file)
 # get the command_ keys
 command_dict = config.options.group_dict("model_command")
 
-jm = mr.JobManager(config.options.redis_url,
-                   config.options.primary_url,
+# initialize the global application settings
+settings.init_redis_connection(config.options.redis_url)
+
+jm = mr.JobManager(config.options.primary_url,
                    config.options.worker_url,
                    config.options.data_dir,
                    command_dict,
